Extracts a few useful helpers to `utils.swift`
